Extracted Death's Embrace functionality to own module
