+++ src/Parser/AfflictionWarlock/CombatLogParser.js
@@ -16,14 +16,15 @@ import UABuffTracker from './Modules/Features/UABuffTracker';
 
 import SoulShardTracker from './Modules/SoulShards/SoulShardTracker';
 import SoulShardDetails from './Modules/SoulShards/SoulShardDetails';
-import DamageDone from "./Modules/WarlockCore/DamageDone";
+import DamageDone from './Modules/WarlockCore/DamageDone';
+import DeathsEmbrace from './Modules/WarlockCore/DeathsEmbrace';
 
 import Haunt from './Modules/Talents/Haunt';
 import MaleficGrasp from './Modules/Talents/MaleficGrasp';
 import Contagion from './Modules/Talents/Contagion';
 import AbsoluteCorruption from './Modules/Talents/AbsoluteCorruption';
 import SoulHarvest from './Modules/Talents/SoulHarvest';
-import DeathsEmbrace from './Modules/Talents/DeathsEmbrace';
+import DeathsEmbraceTalent from './Modules/Talents/DeathsEmbraceTalent';
 import SiphonLifeUptime from './Modules/Talents/SiphonLifeUptime';
 
 import TheMasterHarvester from './Modules/Items/Legendaries/TheMasterHarvester';
@@ -47,6 +48,7 @@ class CombatLogParser extends MainCombatLogParser {
     soulShardTracker: SoulShardTracker,
     soulShardDetails: SoulShardDetails,
     damageDone: DamageDone,
+    deathsEmbrace: DeathsEmbrace,
 
     //Talents
     haunt: Haunt,
@@ -54,7 +56,7 @@ class CombatLogParser extends MainCombatLogParser {
     contagion: Contagion,
     absoluteCorruption: AbsoluteCorruption,
     soulHarvest: SoulHarvest,
-    deathsEmbrace: DeathsEmbrace,
+    deathsEmbraceTalent: DeathsEmbraceTalent,
     siphonLifeUptime: SiphonLifeUptime,
 
     //Legendaries
